Delete Confirm message show, search problem solve

+++ src/components/task/ActionBar.jsx
@@ -1,21 +1,37 @@
+import { toast } from "react-toastify";
 import { useTasksDispatch } from "../../contexts/tasksContext";
 import SearchBox from "./SearchBox";
+import Swal from "sweetalert2";
 
 const ActionBar = ({ showModal }) => {
   const dispatch = useTasksDispatch();
   // All Delete Task
   const handleDeleteAllTask = () => {
-    dispatch({
-      type: "delete-all-task",
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, Delete All!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+         dispatch({
+           type: "delete-all-task",
+         });
+         toast.success("This oparation is successfully.");
+      }
     });
-    toast.success("This oparation is successfully.");
+
+
   };
 
   return (
     <div className="mb-14 items-center justify-between sm:flex">
       <h2 className="text-2xl font-semibold max-sm:mb-4">Your Tasks</h2>
       <div className="flex items-center space-x-5">
-        <SearchBox/>
+        <SearchBox />
         <button
           onClick={showModal}
           className="rounded-md bg-blue-500 px-3.5 py-2.5 text-sm font-semibold"
